+++ cargo-audit/src/commands/audit.rs
@@ -17,6 +17,7 @@ use clap::Parser;
 use rustsec::platforms::target::{Arch, OS};
 use std::{path::PathBuf, process::exit};
 
+#[cfg(feature = "binary-scanning")]
 use self::binary_scanning::BinCommand;
 #[cfg(feature = "fix")]
 use self::fix::FixCommand;
@@ -147,6 +148,8 @@ pub enum AuditSubcommand {
 impl AuditCommand {
     /// Get the color configuration
     pub fn color_config(&self) -> Option<ColorChoice> {
+        // suppress the warning that occurs with the `binary-scanning` feature disabled
+        #[allow(unused_mut)]
         let mut raw_color_setting = self.color.as_ref();
         #[cfg(feature = "binary-scanning")]
         if let Some(AuditSubcommand::Bin(ref command)) = self.subcommand {
@@ -181,6 +184,10 @@ impl From<AuditCommand> for CliConfig {
 
 impl Override<AuditConfig> for AuditCommand {
     fn override_config(&self, config: AuditConfig) -> Result<AuditConfig, FrameworkError> {
+        #[cfg(feature = "binary-scanning")]
+        if let Some(AuditSubcommand::Bin(bin)) = &self.subcommand {
+            return CliConfig::from(bin.clone()).override_config(config);
+        }
         CliConfig::from(self.clone()).override_config(config)
     }
 }