+++ cargo-audit/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.17.2 (2022-10-07)
+### Changed
+
+ - Fixed the screenshot URL in README.md
+
+## 0.17.1 (2022-10-07)
+### Added
+
+ - Initial support for scanning binaries built with [`cargo auditable`](https://github.com/rust-secure-code/cargo-auditable)
+
 ## 0.17.0 (2022-05-23)
 ### Changed
 - Update Abscissa to 0.6; replace `gumdrop` with `clap` v3 ([#525])